[RFC] release notes: reduce root directory clutter and add owners
We periodically need to make updates and clarifying changes to them
and we do have an active set of folks contributing to their curation.
But due to the root level OWNERS file and the filters implementation,
we can't assign the logical reviewer/approvers and the existing
oversubscribed root level reviewer/approvers come to unintentionally
block trivially merged fixups that are meaningful to our end users.

Moving the files to a subdirectory visually simplifies the root of the
repo and allows us to delegate OWNERS.

For the current active releases symlinks are left in the root directory
to reduce the user impact of broken links.

PoC-Not-Yet-Signed-off-by: Tim Pepper <tpepper@vmware.com>